+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 ## [Unreleased]
 ### Added
 - parse method to PixPullSubscription
+- payment and pdf methods to CreditNote, resendToken method to CreditSigner
 
 ### Fixed
 - docstrings for Pix, Issuing and Credit resources to match the business rules stated in the API reference